backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-genai>=1.38.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=3.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]
